Update 220106
Updated Model.dll

+++ backend/BackendProductConfigurator/Model/ProductDependencies.cs
@@ -2,7 +2,7 @@
 {
     public class ProductDependencies
     {
-        public ProductDependencies(int basePrice, List<string> defaultOptions, Dictionary<string, List<string>> replacementGroups, Dictionary<string, List<string>> requirements, Dictionary<string, List<string>> incompabilities, Dictionary<string, List<string>> groupRequirements, Dictionary<string, float> priceList)
+        public ProductDependencies(float basePrice, List<string> defaultOptions, Dictionary<string, List<string>> replacementGroups, Dictionary<string, List<string>> requirements, Dictionary<string, List<string>> incompabilities, Dictionary<string, List<string>> groupRequirements, Dictionary<string, float> priceList)
         {
             BasePrice = basePrice;
             DefaultOptions = defaultOptions;
@@ -11,7 +11,7 @@
             Incompabilities = incompabilities;
             PriceList = priceList;
         }
-        public int BasePrice { get; set; }
+        public float BasePrice { get; set; }
         public List<string> DefaultOptions { get; set; }
         public Dictionary<string, List<string>> ReplacementGroups { get; set; }
         public Dictionary<string, List<string>> Requirements { get; set; }
